# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Connectly social network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  avatarUrl: string | null;
  createdAt: string;
  followerCount: number;
  followingCount: number;
  postCount: number;
  isFollowing: boolean;
  isMe: boolean;
}

export interface UpdateMeBody {
  displayName?: string;
  username?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Post {
  id: number;
  author: User;
  content: string;
  /** @nullable */
  imageUrl: string | null;
  createdAt: string;
  likeCount: number;
  commentCount: number;
  likedByMe: boolean;
}

export interface CreatePostBody {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Comment {
  id: number;
  postId: number;
  author: User;
  content: string;
  createdAt: string;
}

export interface CreateCommentBody {
  content: string;
}

export interface LikeStatus {
  liked: boolean;
  likeCount: number;
}

export interface FollowStatus {
  following: boolean;
  followerCount: number;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  like: "like",
  comment: "comment",
  follow: "follow",
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  actor: User;
  /** @nullable */
  postId: number | null;
  createdAt: string;
}

export type SearchUsersParams = {
  q?: string;
};