# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Connectly social network API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: users
  - name: posts
  - name: comments
  - name: follows
  - name: feed
  - name: notifications
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get the current authenticated user's profile (auto-creates if missing)
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by name or username
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get a user profile by username
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [posts]
      summary: List posts by a user
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [follows]
      summary: Follow a user
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
    delete:
      operationId: unfollowUser
      tags: [follows]
      summary: Unfollow a user
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"

  /users/{username}/followers:
    get:
      operationId: getFollowers
      tags: [follows]
      summary: List followers of a user
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{username}/following:
    get:
      operationId: getFollowing
      tags: [follows]
      summary: List users a user is following
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Personalized feed (posts from people you follow + your own)
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /feed/explore:
    get:
      operationId: getExplore
      tags: [feed]
      summary: Explore feed (recent posts from everyone)
      responses:
        "200":
          description: Explore feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /feed/trending:
    get:
      operationId: getTrending
      tags: [feed]
      summary: Trending posts (most liked in the past week)
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /suggestions:
    get:
      operationId: getSuggestedUsers
      tags: [users]
      summary: Suggested users to follow
      responses:
        "200":
          description: Suggested users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete your own post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatus"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatus"

  /posts/{id}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete your own comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Recent notifications for the current user
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, username, displayName, bio, avatarUrl, createdAt]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        followerCount:
          type: integer
        followingCount:
          type: integer
        postCount:
          type: integer
        isFollowing:
          type: boolean
        isMe:
          type: boolean
      required:
        - id
        - username
        - displayName
        - bio
        - avatarUrl
        - createdAt
        - followerCount
        - followingCount
        - postCount
        - isFollowing
        - isMe

    UpdateMeBody:
      type: object
      properties:
        displayName:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    Post:
      type: object
      properties:
        id:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        likeCount:
          type: integer
        commentCount:
          type: integer
        likedByMe:
          type: boolean
      required:
        - id
        - author
        - content
        - imageUrl
        - createdAt
        - likeCount
        - commentCount
        - likedByMe

    CreatePostBody:
      type: object
      properties:
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
      required: [content]

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, postId, author, content, createdAt]

    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required: [content]

    LikeStatus:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required: [liked, likeCount]

    FollowStatus:
      type: object
      properties:
        following:
          type: boolean
        followerCount:
          type: integer
      required: [following, followerCount]

    Notification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [like, comment, follow]
        actor:
          $ref: "#/components/schemas/User"
        postId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, type, actor, postId, createdAt]