# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Connectly social network API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Comment {
  id: number;
  postId: number;
  author: User;
  content: string;
  createdAt: Date;
}