# lib/api-zod/src/generated/types/followStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Connectly social network API
 * OpenAPI spec version: 0.1.0
 */

export interface FollowStatus {
  following: boolean;
  followerCount: number;
}